requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch
torchvision
fastapi
uvicorn
python-multipart
python-dotenv
numpy
pillow
opencv-python-headless
ultralytics